requirements.txt
mysql-connector == 2.2.9
sys
threading
socket